Modify: Feed 컴포넌트화
