+++ src/pages/hayoung/Main/Comment/Comment.js
@@ -4,12 +4,12 @@ import "./comment.scss";
 const Comment = ({ arr }) => {
   return (
     <ul className="comments">
-      {arr.map(comment => {
+      {arr.map(e => {
         return (
-          <li className="comment" key={comment.id}>
-            <h1 className="commentUserId">{comment.userName}</h1>
-            <h1 className="userComment">{comment.content}</h1>
-            <button className="deleteBtn">삭제{comment.isLiked}</button>
+          <li className="comment" key={e.id}>
+            <h1 className="commentUserId">{e.userName}</h1>
+            <h1 className="userComment">{e.content}</h1>
+            <button className="deleteBtn">삭제{e.isLiked}</button>
           </li>
         );
       })}
