+++ src/pages/hayoung/Main/Feed/Feed.scss
@@ -1,0 +1,104 @@
+.feeds {
+  width: auto;
+  height: auto;
+  margin-top: 60px;
+  background-color: white;
+  border: 1px solid rgba(0, 0, 0, 0.3);
+
+  .feedIdProfile {
+    display: flex;
+    justify-content: flex-start;
+    text-align: center;
+
+    .feedIdPhoto {
+      margin: 10px 0 10px 30px;
+      width: 40px;
+      border-radius: 100%;
+    }
+    .feedIdName {
+      margin-top: 25px;
+      margin-left: 20px;
+      font-weight: bold;
+    }
+    .dot_icon {
+      display: flex;
+      width: 20px;
+      height: 20px;
+      margin: 20px 10px 0 430px;
+    }
+  }
+
+  .feedPhoto {
+    width: 100%;
+    height: 450px;
+  }
+
+  .feed_icons {
+    display: flex;
+    flex-direction: row;
+    margin: 10px;
+
+    img {
+      margin: 0px 7px;
+      width: 25px;
+
+      &:nth-child(4) {
+        margin-left: 465px;
+      }
+    }
+  }
+
+  .like_token {
+    display: flex;
+    justify-content: flex-start;
+    margin: 5px 17px;
+    text-align: center;
+
+    .like_token_profile {
+      width: 20px;
+      height: 20px;
+      border-radius: 100%;
+    }
+    p {
+      margin-left: 5px;
+    }
+  }
+
+  .feed_text {
+    display: flex;
+    flex-direction: row;
+    margin-top: 12px;
+    font-size: 15px;
+
+    .feed_user_id {
+      margin-left: 18px;
+      margin-right: 5px;
+      font-weight: bold;
+    }
+  }
+
+  .comment_form {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    margin-top: 20px;
+    width: 100%;
+    height: 40px;
+    border-top: 1px solid rgba(0, 0, 0, 0.3);
+
+    .comment_input {
+      width: 100%;
+      padding-left: 18px;
+      border: none;
+      font-size: 17px;
+      outline: none;
+    }
+    .comment_btn {
+      width: 10%;
+      border: none;
+      background-color: white;
+      color: #5ba8f2;
+      cursor: pointer;
+    }
+  }
+}
