+++ src/pages/hayoung/Main/Main.scss
@@ -9,259 +9,3 @@ body {
   height: 1000px;
   background-color: #fafafa;
 }
-
-.feeds {
-  width: auto;
-  height: auto;
-  margin-top: 60px;
-  background-color: white;
-  border: 1px solid rgba(0, 0, 0, 0.3);
-
-  .feedIdProfile {
-    display: flex;
-    justify-content: flex-start;
-    text-align: center;
-
-    .feedIdPhoto {
-      margin: 10px 0 10px 30px;
-      width: 40px;
-      border-radius: 100%;
-    }
-    .feedIdName {
-      margin-top: 25px;
-      margin-left: 20px;
-      font-weight: bold;
-    }
-    .dot_icon {
-      display: flex;
-      width: 20px;
-      height: 20px;
-      margin: 20px 10px 0 430px;
-    }
-  }
-
-  .feedPhoto {
-    width: 100%;
-    height: 450px;
-  }
-
-  .feed_icons {
-    display: flex;
-    flex-direction: row;
-    margin: 10px;
-
-    img {
-      margin: 0px 7px;
-      width: 25px;
-
-      &:nth-child(4) {
-        margin-left: 465px;
-      }
-    }
-  }
-
-  .like_token {
-    display: flex;
-    justify-content: flex-start;
-    margin: 5px 17px;
-    text-align: center;
-
-    .like_token_profile {
-      width: 20px;
-      height: 20px;
-      border-radius: 100%;
-    }
-    p {
-      margin-left: 5px;
-    }
-  }
-
-  .feed_text {
-    display: flex;
-    flex-direction: row;
-    margin-top: 12px;
-    font-size: 15px;
-
-    .feed_user_id {
-      margin-left: 18px;
-      margin-right: 5px;
-      font-weight: bold;
-    }
-  }
-
-  .comment_form {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: 20px;
-    width: 100%;
-    height: 40px;
-    border-top: 1px solid rgba(0, 0, 0, 0.3);
-
-    .comment_input {
-      width: 100%;
-      padding-left: 18px;
-      border: none;
-      font-size: 17px;
-      outline: none;
-    }
-    .comment_btn {
-      width: 10%;
-      border: none;
-      background-color: white;
-      color: #5ba8f2;
-      cursor: pointer;
-    }
-  }
-}
-
-.main_right {
-  justify-content: flex-end;
-  margin-top: 50px;
-  margin-left: 30px;
-
-  .user_id {
-    display: flex;
-
-    .user_profile {
-      width: 50px;
-      height: 50px;
-      border-radius: 100%;
-    }
-    .user_id_name {
-      display: flex;
-      flex-direction: column;
-      margin-top: 10px;
-      margin-left: 10px;
-
-      p {
-        &:nth-child(1) {
-          font-weight: bold;
-        }
-
-        &:nth-child(2) {
-          color: gray;
-        }
-      }
-    }
-  }
-
-  .story_box {
-    justify-content: flex-end;
-    width: 300px;
-    height: 230px;
-    margin-top: 16px;
-    border: 1px solid rgba(0, 0, 0, 0.3);
-    background-color: white;
-
-    .box_text {
-      display: flex;
-      justify-content: space-between;
-      margin: 15px;
-
-      p {
-        &:nth-child(1) {
-          color: gray;
-        }
-        &:nth-child(2) {
-          font-weight: bold;
-        }
-      }
-    }
-
-    .story_id {
-      display: flex;
-
-      .story_id_profile {
-        width: 40px;
-        height: 40px;
-        margin-left: 15px;
-        margin-bottom: 15px;
-        padding: 2px;
-        border-radius: 100%;
-        outline: 2px solid #ff0000;
-      }
-      .story_id_name {
-        display: flex;
-        flex-direction: column;
-        margin-top: 10px;
-        margin-left: 10px;
-
-        p {
-          &:nth-child(1) {
-            font-weight: bold;
-          }
-
-          &:nth-child(2) {
-            color: gray;
-          }
-        }
-      }
-    }
-  }
-
-  .recommend_box {
-    justify-content: flex-end;
-    width: 300px;
-    height: 200px;
-    margin-top: 16px;
-    border: 1px solid rgba(0, 0, 0, 0.3);
-    background-color: white;
-
-    .box_text {
-      display: flex;
-      justify-content: space-between;
-      margin: 15px;
-
-      p {
-        &:nth-child(1) {
-          color: gray;
-        }
-        &:nth-child(2) {
-          font-weight: bold;
-        }
-      }
-    }
-
-    .recommend_id {
-      display: flex;
-      justify-content: space-around;
-      width: 100%;
-
-      .recommend_id_profile {
-        width: 40px;
-        height: 40px;
-        // margin-left: 15px;
-        margin-bottom: 15px;
-        border-radius: 100%;
-      }
-      .recommend_id_name {
-        p {
-          &:nth-child(1) {
-            font-weight: bold;
-          }
-
-          &:nth-child(2) {
-            color: gray;
-          }
-        }
-      }
-
-      .follower_text {
-        display: flex;
-        justify-content: flex-end;
-        width: 60px;
-        color: #5ba8f2;
-        font-weight: bold;
-        text-align: center;
-      }
-    }
-  }
-
-  .info {
-    margin-top: 20px;
-    color: gray;
-    opacity: 0.3;
-    margin-bottom: 20px;
-  }
-}
