+++ src/pages/hayoung/Main/MainRight/MainRight.scss
@@ -1,0 +1,150 @@
+.main_right {
+  justify-content: flex-end;
+  margin-top: 50px;
+  margin-left: 30px;
+
+  .user_id {
+    display: flex;
+
+    .user_profile {
+      width: 50px;
+      height: 50px;
+      border-radius: 100%;
+    }
+    .user_id_name {
+      display: flex;
+      flex-direction: column;
+      margin-top: 10px;
+      margin-left: 10px;
+
+      p {
+        &:nth-child(1) {
+          font-weight: bold;
+        }
+
+        &:nth-child(2) {
+          color: gray;
+        }
+      }
+    }
+  }
+
+  .story_box {
+    justify-content: flex-end;
+    width: 300px;
+    height: 230px;
+    margin-top: 16px;
+    border: 1px solid rgba(0, 0, 0, 0.3);
+    background-color: white;
+
+    .box_text {
+      display: flex;
+      justify-content: space-between;
+      margin: 15px;
+
+      p {
+        &:nth-child(1) {
+          color: gray;
+        }
+        &:nth-child(2) {
+          font-weight: bold;
+        }
+      }
+    }
+
+    .story_id {
+      display: flex;
+
+      .story_id_profile {
+        width: 40px;
+        height: 40px;
+        margin-left: 15px;
+        margin-bottom: 15px;
+        padding: 2px;
+        border-radius: 100%;
+        outline: 2px solid #ff0000;
+      }
+      .story_id_name {
+        display: flex;
+        flex-direction: column;
+        margin-top: 10px;
+        margin-left: 10px;
+
+        p {
+          &:nth-child(1) {
+            font-weight: bold;
+          }
+
+          &:nth-child(2) {
+            color: gray;
+          }
+        }
+      }
+    }
+  }
+
+  .recommend_box {
+    justify-content: flex-end;
+    width: 300px;
+    height: 200px;
+    margin-top: 16px;
+    border: 1px solid rgba(0, 0, 0, 0.3);
+    background-color: white;
+
+    .box_text {
+      display: flex;
+      justify-content: space-between;
+      margin: 15px;
+
+      p {
+        &:nth-child(1) {
+          color: gray;
+        }
+        &:nth-child(2) {
+          font-weight: bold;
+        }
+      }
+    }
+
+    .recommend_id {
+      display: flex;
+      justify-content: space-around;
+      width: 100%;
+
+      .recommend_id_profile {
+        width: 40px;
+        height: 40px;
+        // margin-left: 15px;
+        margin-bottom: 15px;
+        border-radius: 100%;
+      }
+      .recommend_id_name {
+        p {
+          &:nth-child(1) {
+            font-weight: bold;
+          }
+
+          &:nth-child(2) {
+            color: gray;
+          }
+        }
+      }
+
+      .follower_text {
+        display: flex;
+        justify-content: flex-end;
+        width: 60px;
+        color: #5ba8f2;
+        font-weight: bold;
+        text-align: center;
+      }
+    }
+  }
+
+  .info {
+    margin-top: 20px;
+    color: gray;
+    opacity: 0.3;
+    margin-bottom: 20px;
+  }
+}
